+++ src/BugsnagUnity/Bugsnag.cs
@@ -75,5 +75,27 @@ namespace BugsnagUnity
     /// </summary>
     /// <param name="inFocus"></param>
     public static void SetApplicationState(bool inFocus) => Client.SetApplicationState(inFocus);
+
+    /// <summary>
+    /// Bugsnag uses the concept of contexts to help display and group your errors.
+    /// Contexts represent what was happening in your game at the time an error
+    /// occurs. By default, this will be set to be your currently active Unity Scene.
+    /// </summary>
+    /// <param name="context"></param>
+    public static void SetContext(string context)
+    {
+      Client.SetContext(context);
+    }
+
+    /// <summary>
+    /// By default, we will automatically notify Bugsnag of any fatal errors (crashes) in your game.
+    /// If you want to stop this from happening, you can set the AutoNotify property to false. It
+    /// is recommended that you set this value by Configuration rather than this method.
+    /// </summary>
+    /// <param name="autoNotify"></param>
+    public static void SetAutoNotify(bool autoNotify)
+    {
+      Client.SetAutoNotify(autoNotify);
+    }
   }
 }