+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Replace Bugsnag.init with Bugsnag.start
   [#227](https://github.com/bugsnag/bugsnag-unity/pull/227)
 
+* Add Bugsnag.SetContext and Bugsnag.SetAutoNotify
+  [#229](https://github.com/bugsnag/bugsnag-unity/pull/229)
+
 ## 4.8.8 (2021-04-21)
 
 ### Bug fixes